pie graph + table (almost done)

+++ app/src/main/java/sfu/cmpt276/carbontracker/Route.java
@@ -7,8 +7,8 @@ interface RouteListener {
 public class Route {
     private boolean isActive;
     private String routeName;
-    private int routeDistanceCity;
-    private int routeDistanceHighway;
+    private double routeDistanceCity;
+    private double routeDistanceHighway;
 
     public Route() {
         isActive = false;
@@ -17,13 +17,13 @@ public class Route {
         routeDistanceHighway = 0;
     }
 
-    public Route(int distanceC, int distanceH){
+    public Route(double distanceC, double distanceH) {
         isActive = false;
         routeDistanceCity = distanceC;
         routeDistanceHighway = distanceH;
     }
 
-    public Route(String name, int distanceC, int distanceH){
+    public Route(String name, double distanceC, double distanceH) {
         isActive = false;
         routeName = name;
         routeDistanceCity = distanceC;
@@ -32,39 +32,39 @@ public class Route {
     }
 
     //return the name
-    public String getRouteName(){
+    public String getRouteName() {
         return routeName;
     }
 
     //set the name
-    public void setRouteName(String name){
-        if(name==null){
+    public void setRouteName(String name) {
+        if (name == null) {
             throw new IllegalArgumentException();
         }
         routeName = name;
     }
 
     //return the distance of city
-    public int getRouteDistanceCity(){
+    public double getRouteDistanceCity() {
         return routeDistanceCity;
     }
 
     //return the distance of highway
-    public int getRouteDistanceHighway(){
+    public double getRouteDistanceHighway() {
         return routeDistanceHighway;
     }
 
     //set the distance of city
-    public void setRouteDistanceCity(int distanceC){
-        if(distanceC<0){
+    public void setRouteDistanceCity(double distanceC) {
+        if (distanceC < 0) {
             throw new IllegalArgumentException();
         }
         routeDistanceCity = distanceC;
     }
 
     //set the distance of highway
-    public void setRouteDistanceHighway(int distanceH){
-        if(distanceH<0){
+    public void setRouteDistanceHighway(double distanceH) {
+        if (distanceH < 0) {
             throw new IllegalArgumentException();
 
         }
